// src/models/CompanyUserProfileExtSkillModel.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { CompanyUserProfileExtSkillTranslationModel } from './CompanyUserProfileExtSkillTranslationModel';
import type { Link } from './Link';
export type CompanyUserProfileExtSkillModel = {
    profileId?: number | null;
    translations?: Array<CompanyUserProfileExtSkillTranslationModel> | null;
    companyId?: number | null;
    companyUserId?: number | null;
    id?: number | null;
    url?: string | null;
    links?: Array<Link> | null;
};